Fix animation clipping with small windows (+ refactor for new filesystem)
