+++ components/Animation/utils.jsx
@@ -1,0 +1,159 @@
+
+const PointSet1 = () => {
+
+    const path1 = [
+        [2599.298864957342, -844.3636519746029],
+        [2690.3746029830672, -820.1607875000084],
+        [2295.7699583085746, 645.2217897637374],
+        [2473.6820111172765, 542.4129187451488],
+        [2736.783243232903, -430.9702762031668],
+        [2826.1970036109656, -404.265288728904],
+        [2589.2624421742867, 475.6223711077539],
+        [2729.749072653589, 394.4260643912444],
+        [2856.338389377531, -75.63615183163802],
+        [2946.915493325826, -49.6129957441799],
+        [2810.5370612051456, 456.8580909151324],
+        [2262.537459012803, 773.5368049663193],
+        [1490.8584569063153, 568.4558973688372],
+        [841.8330237966927, -554.6530872407102],
+        [1086.0964735288117, -1655.924777005776],
+        [1178.1531464725228, -1635.3601558008038],
+        [1114.276961033294, -1346.8810830947004],
+        [2008.8927285572258, 201.21463968544094],
+        [2297.115833490085, 277.812597454762],
+        [2599.298864957342, -844.3636519746029],
+    ];
+
+    const path2 = [
+        [2195.6552485554785, 658.2504280081873],
+        [2221.0454315605134, 560.2874206241852],
+        [1704.7602894770473, 423.07674232553313],
+        [993.6375515179134, -807.4912463339411],
+        [940.7001543158904, -570.306055731196],
+        [1551.7751059659217, 487.1349526935969],
+        [2195.6552485554785, 658.2504280081873],
+    ];
+
+    const path3 = [
+        [2245.553232676324, 469.2924542063363],
+        [2272.6100515618296, 368.8055430764441],
+        [1947.9683215700952, 282.5278326086884],
+        [1084.6408219734526, -1211.4255791076862],
+        [1024.4432621169926, -943.6208083455064],
+        [1767.5197601043553, 342.24584342931576],
+        [2245.553232676324, 469.2924542063363],
+    ];
+
+    return [path1, path2, path3];
+};
+
+
+const PointSet2 = () => {
+
+    const path1 = [
+        [1131.8620152154367, -1221.5767428556303],
+        [1222.1626943370204, -1286.7628272354673],
+        [2267.1056694702397, 170.89387949308912],
+        [2306.2291826284268, -68.77480197125513],
+        [1612.8605130372384, -1037.9263744846862],
+        [1704.2058218456398, -1099.7196981001778],
+        [2331.6449733185846, -224.47601051575876],
+        [2362.525493696187, -413.7388417657596],
+        [2027.3376357677619, -881.3309124978762],
+        [2118.930008103658, -944.6985664189737],
+        [2480.0918960913014, -440.9035071543734],
+        [2359.593716497068, 297.32440972468453],
+        [1594.4711679450165, 849.6333534017821],
+        [81.49355093991245, 602.6775157600621],
+        [-742.9441779041043, -444.9604916574872],
+        [-655.2369693587771, -513.7667468387103],
+        [-439.1933354233256, -239.4366086428082],
+        [1646.2964483174871, 100.96985169387051],
+        [1932.0706852690698, -105.31827253125479],
+        [1131.8620152154367, -1221.5767428556303],
+    ];
+
+    const path2 = [
+        [2204.0928542193515, 272.2125361061708],
+        [2133.496117350925, 175.67124373591102],
+        [1621.59470012397, 545.18644896871],
+        [-36.143472935264214, 274.6005994603415],
+        [141.17235398474742, 500.53579304518286],
+        [1565.6848880584935, 733.0540738257099],
+        [2204.0928542193515, 272.2125361061708],
+    ];
+
+    const path3 = [
+        [2068.6121335029047, 85.15211354008466],
+        [1996.9543320695052, -14.802501569031136],
+        [1675.0698319548706, 217.55042521080526],
+        [-337.48277140352627, -110.95122802124123],
+        [-137.59520248461058, 144.54101010715203],
+        [1594.6358577711624, 427.28844191229496],
+        [2068.6121335029047, 85.15211354008466],
+    ];
+
+    return [path1, path2, path3];
+};
+
+const PointSet3 = () => {
+
+    const path1 = [
+        [1511.5917743472655, -1004.2590745605789],
+        [1615.0221769967784, -1023.6427358595708],
+        [1921.6712064284177, 643.0163675547201],
+        [2051.5735135407845, 453.87534926114535],
+        [1848.8734353134316, -653.6793148101153],
+        [1951.8314207142116, -669.7421723654782],
+        [2135.9640447283336, 330.999180529709],
+        [2238.533382451991, 181.63136825986527],
+        [2140.1761507165493, -353.0001707144358],
+        [2243.978946485936, -370.3094500494616],
+        [2349.970520118017, 205.7245542354019],
+        [1949.8538598186624, 788.3211495436742],
+        [1073.4890051604527, 952.5460429308325],
+        [-120.52727847616795, 132.51962696908328],
+        [-403.68741921325926, -1094.8808881252785],
+        [-301.0171405558742, -1118.4031038485966],
+        [-226.73339832096667, -796.9367915434569],
+        [1419.0988553164239, 333.3903643554854],
+        [1746.4222488509852, 272.052127800603],
+        [1511.5917743472655, -1004.2590745605789],
+    ];
+
+    const path2 = [
+        [1827.0266480503458, 704.2710602436306],
+        [1805.5291448420733, 593.3281563648001],
+        [1219.203362622194, 703.1982961577642],
+        [-89.05490840656651, -195.28887948772993],
+        [-28.40163176229641, 69.21951574033571],
+        [1095.7986389254515, 841.3001391746628],
+        [1827.0266480503458, 704.2710602436306],
+    ];
+
+    const path3 = [
+        [1786.4916589452218, 489.8325883362648],
+        [1765.4607985655582, 375.544688459108],
+        [1396.7776761223924, 444.63218406749354],
+        [-191.49364352393897, -646.1636933681084],
+        [-123.43983908681597, -347.3035581999942],
+        [1243.605958429558, 591.5605133534826],
+        [1786.4916589452218, 489.8325883362648],
+    ];
+
+    return [path1, path2, path3];
+};
+
+const PointSetChooser = (x) => {
+    if (x === "1") {
+        return PointSet1();
+    } else if (x === "2") {
+        return PointSet2();
+    } else if (x === "3") {
+        return PointSet3();
+    }
+};
+
+
+export default PointSetChooser;
+
