+++ layout/Challenges/components/Hero/index.jsx
@@ -2,12 +2,12 @@ import dynamic from "next/dynamic";
 
 import Title from "./Title";
 
-const Animation = dynamic(() => import("./Animation"), { ssr: false });
+const Animation = dynamic(() => import("@components/Animation"), { ssr: false });
 export default function Hero() {
   return (
     <div className=" bg-secondary">
       <div className="spacing relative z-20 md:pb-32">
-        <Animation />
+        <Animation type="2"/>
         <div className="pt-2">
           <Title />
         </div>
