+++ layout/Home/components/Hero/index.jsx
@@ -11,14 +11,14 @@ import Organization from "./Organization";
 import ExpectList from "./ExpectList";
 import Pitch from "./Pitch";
 
-const Animation = dynamic(() => import("../Animation"), { ssr: false });
+const Animation = dynamic(() => import("@components/Animation"), { ssr: false });
 
 export default function Hero() {
   const { isAuthenticated } = useAuth();
   return (
     <div className="bg-secondary">
       <div className="spacing relative z-40 pb-32">
-        <Animation />
+        <Animation type="1"/>
         <div className="pt-2">
           <Title />
         </div>
