+++ layout/Schedule/components/Hero/index.jsx
@@ -2,13 +2,13 @@ import dynamic from "next/dynamic";
 
 import Title from "./Title";
 
-const Animation = dynamic(() => import("./Animation"), { ssr: false });
+const Animation = dynamic(() => import("@components/Animation"), { ssr: false });
 
 export default function Hero() {
   return (
     <div className="bg-tertiary">
       <div className="spacing relative z-20 pb-32">
-        <Animation />
+        <Animation type="3"/>
         <div className="pt-2">
           <Title />
         </div>
