+++ layout/Team/components/Hero/index.jsx
@@ -3,13 +3,13 @@ import dynamic from "next/dynamic";
 import Title from "./Title";
 import Organization from "./Organization";
 
-const Animation = dynamic(() => import("./Animation"), { ssr: false });
+const Animation = dynamic(() => import("@components/Animation"), { ssr: false });
 
 export default function Hero() {
   return (
     <div className="spacing bg-primary">
       <div className="relative z-50">
-        <Animation />
+        <Animation type="1"/>
         <div className="pt-2">
           <Title />
         </div>
